move edition form to components dir
